backend_python/requirements.txt
# TikTok Live Integration Dependencies
TikTokLive==5.0.8
pyee==11.1.0
websocket-client==1.6.4
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0

